rd-143 refactor contexts
