+++ packages/robindoc/src/contexts/anchor-context.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+export const RegisterContext = createContext<((ref: HTMLHeadingElement) => void) | null>(null);
+export const CurrentSectionContext = createContext<number | null>(null);
